some gamepad functionality moved from windows to global
