+++ src/gamepad/handler_windows.go
@@ -1,20 +1,10 @@
 package gamepad
 
 import (
-	"math"
 	"time"
 
 	"github.com/pion/webrtc/v3"
 	"github.com/pquerna/ffjson/ffjson"
-)
-
-const (
-	threshold float64 = 1e-9
-)
-
-var (
-	prevThumbLY float64
-	prevThumbRY float64
 )
 
 var buttonValueToHexMap = map[int]uint16{
@@ -109,28 +99,6 @@ func convertGamepadButtons(buttons [16]gamepadButton) Button {
 		}
 	}
 	return result
-}
-
-func fixLYAxis(value float64) float64 {
-
-	if math.Abs(value-prevThumbLY) <= threshold {
-		return prevThumbLY
-	}
-
-	prevThumbLY = -value
-	return -value
-
-}
-
-func fixRYAxis(value float64) float64 {
-
-	if math.Abs(value-prevThumbRY) <= threshold {
-		return prevThumbRY
-	}
-
-	prevThumbRY = -value
-	return -value
-
 }
 
 func convertFloatToUint8(value float64) uint8 {
